feat(types): add support for continuous logs such as container logs
Includes some sweep up refactoring
