+++ package.json
@@ -12,15 +12,6 @@
     "selection",
     "app launcher"
   ],
-  "bin": {
-    "yall": "index.js"
-  },
-  "dependencies": {
-    "chalk": "^5.2.0",
-    "commander": "^10.0.0",
-    "inquirer": "^9.1.4",
-    "inquirer-autocomplete-prompt": "^3.0.0"
-  },
   "description": "A CLI tool to launch applications in half the time",
   "repository": {
     "type": "git",
@@ -30,9 +21,22 @@
     "url": "https://github.com/Amber-Williams/yall/issues"
   },
   "homepage": "https://github.com/Amber-Williams/yall#readme",
+  "bin": {
+    "yall": "index.js"
+  },
+  "dependencies": {
+    "chalk": "^5.2.0",
+    "commander": "^10.0.0",
+    "inquirer": "^9.1.4",
+    "inquirer-autocomplete-prompt": "^3.0.0"
+  },
   "devDependencies": {
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
     "semantic-release": "^21.0.1"
+  },
+  "engines": {
+    "npm": ">=8.0.0 <9.0.0",
+    "node": ">=18"
   }
 }
